+++ lib/ui/messages/MessagesConversationPanel.dart
@@ -5,6 +5,7 @@ import 'package:collection/collection.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:http/http.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:neom/ext/Social.dart';
 import 'package:neom/service/Analytics.dart';
@@ -32,9 +33,14 @@ import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/social.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
+import 'package:sprintf/sprintf.dart';
 import 'package:universal_io/io.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:permission_handler/permission_handler.dart';
+
+import 'package:neom/platform_impl/stub.dart'
+  if (dart.library.io) 'package:neom/platform_impl/mobile.dart'
+  if (dart.library.html) 'package:neom/platform_impl/web.dart';
 
 enum FileType { image, video, audio, file }
 
@@ -63,6 +69,7 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   final GlobalKey _targetMessageContentItemKey = GlobalKey();
   final GlobalKey _inputFieldKey = GlobalKey();
   final FocusNode _inputFieldFocus = FocusNode();
+  final FileHelper _fileHelper = FileHelper();
 
   _ScrollTarget? _shouldScrollToTarget;
   Map<String, Uint8List?> _userPhotosCache = {};
@@ -71,6 +78,7 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   bool _loading = false;
   bool _loadingMore = false;
   bool _submitting = false;
+  Map<String, bool> _uploadingFiles = {};
 
   // Use the actual Auth2 accountId instead of a placeholder.
   String? get _currentUserId => Auth2().accountId;
@@ -108,6 +116,8 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
 
     // Load conversation (if needed) and messages from the backend
     _initConversationAndMessages();
+
+    _fileHelper.initializePicker();
 
     super.initState();
   }
@@ -940,7 +950,9 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
         return _buildAudioAttachment(context, file, type, inMessage: false);
       }
       textStyleKey = 'widget.title.small';
-      onTap = () => _removeAttachedFiles([file]);
+      if (_uploadingFiles[name] != true) {
+        onTap = () => _removeAttachedFiles([file]);
+      }
       if (file is PlatformFile) {
         name = file.name;
         extension = file.extension;
@@ -1016,7 +1028,8 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   Widget _buildAttachmentContainer({required Widget child,
     bool inMessage = true, void Function()? onTap,
     void Function()? onRemove}) {
-    return Column(
+    return Stack(
+      alignment: Alignment.center,
       children: [
         GestureDetector(
           onTap: onTap,
@@ -1037,6 +1050,10 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
             ),
           ),
         ),
+        Visibility(
+          visible: !inMessage && (_uploadingFiles[name] == true),
+          child: CircularProgressIndicator(strokeWidth: 3, valueColor: AlwaysStoppedAnimation<Color?>(Styles().colors.fillColorSecondary),),
+        ),
       ],
     );
   }
@@ -1044,8 +1061,15 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   Widget _buildAttachedMediaEntry(BuildContext context, dynamic file, FileType type,
       {bool inMessage = true}) {
     String? path, url;
+    Uint8List? data;
     if (file is FileAttachment) {
-      url = file.url;
+      if (file.url != null) {
+        url = file.url;
+      } else if (file.data != null) {
+        data = file.data;
+      }
+    } else if (kIsWeb && file is PlatformFile) {
+      data = file.bytes;
     } else {
       String? filePath = _getFilePath(file);
       if (kIsWeb) {
@@ -1054,12 +1078,17 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
         path = filePath;
       }
     }
-    if (path == null && url == null) {
+    if (path == null && url == null && data == null) {
       return const SizedBox();
     }
     Widget? widget;
     if (type == FileType.image) {
-      if (kIsWeb || file is FileAttachment) {
+      if (data != null) {
+        widget = Image.memory(data, fit: BoxFit.cover,
+          errorBuilder: (BuildContext context, Object error, StackTrace? stackTrace) =>
+          _imageErrorBuilder,
+        );
+      } else if (kIsWeb || file is FileAttachment) {
         widget = Image.network(url ?? '', fit: BoxFit.cover,
           errorBuilder: (BuildContext context, Object error, StackTrace? stackTrace) =>
           _imageErrorBuilder,
@@ -1069,6 +1098,8 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
           errorBuilder: (BuildContext context, Object error, StackTrace? stackTrace) =>
           _imageErrorBuilder,
         );
+      } else {
+        return const SizedBox();
       }
     }
     else if (type == FileType.video) {
@@ -1228,15 +1259,19 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
       extendLimitToMessageId: extendLimitToMessageId,
     );
     if (loadAttachmentUrls) {
+      List<String> fileIds = [];
       for (Message message in loadedMessages ?? []) {
-        List<String>? filenames = message.fileAttachments?.where((e) => e.type != FileType.file).map((e) => e.name).whereNotNull().toList();
-        if (filenames != null && filenames.isNotEmpty) {
-          Map<String, String>? urls = await Content().getFileContentDownloadUrls(filenames, Content.conversationsContentCategory, entityId: '$_conversationId/${message.globalId}');
-          if (urls != null && urls.isNotEmpty) {
+        fileIds.addAll(message.fileAttachments?.where((e) => e.type != FileType.file).map((e) => e.id).whereNotNull() ?? []);
+      }
+
+      if (CollectionUtils.isNotEmpty(fileIds)) {
+        List<FileContentItemReference>? fileRefs = await Content().getFileContentDownloadUrls(fileIds, Content.conversationsContentCategory, entityId: _conversationId);
+        if (CollectionUtils.isNotEmpty(fileRefs)) {
+          for (Message message in loadedMessages ?? []) {
             for (FileAttachment file in message.fileAttachments ?? []) {
-              for (MapEntry<String, String> urlEntry in urls.entries) {
-                if (urlEntry.value == file.name) {
-                  file.url = urlEntry.key;
+              for (FileContentItemReference ref in fileRefs ?? []) {
+                if (ref.id == file.id) {
+                  file.url = ref.url;
                 }
               }
             }
@@ -1272,8 +1307,13 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
     if (StringUtils.isNotEmpty(messageText) && _conversationId != null && _currentUserId != null && !_submitting) {
       FocusScope.of(context).requestFocus(FocusNode());
 
+      List<FileContentItemReference>? fileRefs;
+      if (_attachedFiles.isNotEmpty) {
+        fileRefs = await _uploadAttachedFiles();
+      }
+
       _inputController.text = '';
-      List<FileAttachment> fileAttachments = _messageFileAttachments;
+      List<FileAttachment> fileAttachments = await _getMessageFileAttachments(fileRefs);
 
       // Create a temporary message and add it immediately
       Message tempMessage = Message(
@@ -1290,6 +1330,8 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
         _shouldScrollToTarget = _ScrollTarget.bottom; //TBD
       });
 
+      _removeAttachedFiles(fileAttachments);
+
       // Send to the backend
       List<Message>? newMessages = await Social().createConversationMessage(
         conversationId: _conversationId!,
@@ -1297,12 +1339,10 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
         fileAttachments: fileAttachments,
       );
 
-      String? messageId;
       if (mounted) {
         if (newMessages != null && newMessages.isNotEmpty) {
           setState(() {
-            Message serverMessage = newMessages.first;
-            messageId = serverMessage.globalId;
+            Message serverMessage = Message.fromOther(newMessages.first, fileAttachments: fileAttachments);
             // Update the temporary message with the server's message if needed
             int index = _messages.indexOf(tempMessage);
             if (index >= 0) {
@@ -1320,17 +1360,35 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
           AppToast.showMessage(Localization().getStringEx('', 'Failed to send message'));
         }
       }
-
-      if (_attachedFiles.isNotEmpty && messageId != null) {
-        _uploadAttachedFiles(messageId);
-      }
     }
   }
 
-  Future<void> _uploadAttachedFiles(String? messageId) async {
-    List<String>? uploaded = await Content().uploadFileContentItems(_attachedFileData, Content.conversationsContentCategory, entityId: '$_conversationId/$messageId');
-    _removeAttachedFiles(uploaded);
-    setStateIfMounted(() {});
+  Future<List<FileContentItemReference>?> _uploadAttachedFiles() async {
+    List<FileContentItemReference>? uploaded = await Content().uploadFileContentItems(_attachedFileData, Content.conversationsContentCategory,
+      entityId: _conversationId, preUpload: _preUploadFile, postUpload: _postUploadFile);
+
+    int failedFileCount = _attachedFileData.length - (uploaded?.length ?? 0);
+    if (failedFileCount > 0) {
+      AppToast.showMessage(sprintf(Localization().getStringEx('', 'Failed to upload %s file(s)'), [failedFileCount]));
+    }
+
+    return uploaded;
+  }
+
+  void _preUploadFile(FileContentItemReference ref) {
+    if (ref.name != null) {
+      setStateIfMounted(() {
+        _uploadingFiles[ref.name!] = true;
+      });
+    }
+  }
+
+  void _postUploadFile(FileContentItemReference ref, Response? response) {
+    if (ref.name != null) {
+      setStateIfMounted(() {
+        _uploadingFiles[ref.name!] = false;
+      });
+    }
   }
 
   Future<void> _updateEditingMessage(String newText) async {
@@ -1439,9 +1497,9 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
     }
     List<dynamic> processed = [];
     for (dynamic file in files) {
-      if (file is String) {
+      if (file is FileAttachment) {
         dynamic found = _attachedFiles.firstWhereOrNull((e) {
-          return _getFileName(e) == file;
+          return _getFileName(e) == file.name;
         });
         if (found != null) {
           processed.add(found);
@@ -1461,16 +1519,11 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   Future<void> _onTapDownloadFile(FileAttachment file, String messageId) async {
     //TODO: implement opening files based on type
     if (StringUtils.isNotEmpty(file.name)) {
-      Map<String, Uint8List> files = await Content().getFileContentItems([file.name!], Content.conversationsContentCategory, entityId: '$_conversationId/$messageId');
-      // if (await _requestStoragePermissions() && files.isNotEmpty) {
-      Uint8List? data = files[file.name];
+      Map<String, Uint8List> files = await Content().getFileContentItems([file.id!], Content.conversationsContentCategory, entityId: _conversationId);
+      Uint8List? data = files[file.id];
       if (CollectionUtils.isNotEmpty(data)) {
         AppFile.downloadFile(context: context, fileName: file.name ?? 'file.out', fileBytes: data);
-        // bool success = await RokwirePlugin.saveDownloadedFile(file.name!, data!);
-        // String message = success ? Localization().getStringEx('', 'File saved') : Localization().getStringEx('', 'Failed to save file');
-        // AppToast.showMessage(message);
       }
-      // }
     }
   }
 
@@ -1517,14 +1570,6 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
       }
     }
   }
-
-  double get _chatBarHeight {
-    RenderObject? chatBarRenderBox = _chatBarKey.currentContext?.findRenderObject();
-    double? chatBarHeight = ((chatBarRenderBox is RenderBox) && chatBarRenderBox.hasSize) ? chatBarRenderBox.size.height : null;
-    return chatBarHeight ?? 0;
-  }
-
-  double get _scrollContentPaddingBottom => _chatBarHeight;
 
   Future<bool> get _checkKeyboardVisible async {
     final checkPosition = () => (MediaQuery.of(context).viewInsets.bottom);
@@ -1621,11 +1666,13 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
     return FileType.values.firstWhereOrNull((e) => e.name == type) ?? FileType.file;
   }
 
-  List<FileAttachment> get _messageFileAttachments =>
-    List.generate(_attachedFiles.length, (index) {
+  List<FileAttachment> _getMessageFileAttachments(List<FileContentItemReference>? fileRefs) {
+    return fileRefs != null ? List.generate(_attachedFiles.length, (index) {
       dynamic file = _attachedFiles.elementAt(index);
       FileType type = _getFileType(file);
       String? name = _getFileName(file);
-      return FileAttachment(name: name, type: type.name);
-    });
+      FileContentItemReference ref = fileRefs.firstWhere((ref) => ref.name == name, orElse: () => FileContentItemReference());
+      return FileAttachment(name: name, type: type.name, id: ref.id, data: ref.data);
+    }) : [];
+  }
 }